+++ releasenotes.md
@@ -1,12 +1,19 @@
 Release Notes
 =============
 
+2.4.0
+------------------
+
+* Added support for the `/email/domain/first_seen` output in minFraud Insights
+  and Factors. This is available as the `FirstSeen` property on
+  `MaxMind.MinFraud.Response.EmailDomain`.
+
 2.3.0 (2020-02-21)
 ------------------
 
 * Added support for the `/email/is_disposable` output in minFraud Insights
   and Factors. This is available as the `IsDisposable` property on
-  `MaxMind.MinFraud.Response`.
+  `MaxMind.MinFraud.Response.Email`.
 * Added the following new values to the `PaymentProcessor` enum:
   * `Cetelem`
   * `Ecommpay`